requirements.txt
discord.py
pytz
matplotlib
seaborn
supabase
python-dotenv